chore: rename __fixture__/ to __fixtures__/ to match beachball ignorePatterns
beachball.config.js#ignorePatterns lists **/__fixtures__/** (plural).
Renaming the directory lets later edits to test fixtures (e.g. adding
the assets field) skip the change-file requirement automatically,
without relying on glob brace expansion which is matcher-dependent.
