+++ packages/monosize/tsconfig.lib.json
@@ -6,5 +6,5 @@
     "types": ["environment"]
   },
   "include": ["src/**/*.mts"],
-  "exclude": ["src/**/*.test.mts", "src/**/__fixture__/", "src/**/__mocks__/"]
+  "exclude": ["src/**/*.test.mts", "src/**/__fixtures__/", "src/**/__mocks__/"]
 }
